+++ nba_stats/nba_api.py
@@ -1,0 +1,70 @@
+import datetime
+
+from requests import get
+import pandas as pd
+
+# Constants
+TODAY = datetime.today()
+BASE_URL = 'http://stats.nba.com/stats/{endpoint}'
+HEADERS = {
+        'Host': 'stats.nba.com',
+        'Connection': 'keep-alive',
+        'Accept': 'application/json, text/plain, */*',
+        'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/78.0.3904.97 Safari/537.36',
+        'Referer': 'https://stats.nba.com/',
+        'Accept-Encoding': 'gzip, deflate, br',
+        'Accept-Language': 'en-US,en;q=0.9',
+}
+
+
+def _api_scrape(json_inp, ndx):
+    """
+    Internal method to streamline the getting of data from the json
+
+    Args:
+        json_inp (json): json input from our caller
+        ndx (int): index where the data is located in the api
+
+    Returns:
+        If pandas is present:
+            DataFrame (pandas.DataFrame): data set from ndx within the
+            API's json
+        else:
+            A dictionary of both headers and values from the page
+    """
+
+    try:
+        headers = json_inp['resultSets'][ndx]['headers']
+        values = json_inp['resultSets'][ndx]['rowSet']
+    except KeyError:
+        # This is so ugly but this is what you get when your data comes out
+        # in not a standard format
+        try:
+            headers = json_inp['resultSet'][ndx]['headers']
+            values = json_inp['resultSet'][ndx]['rowSet']
+        except KeyError:
+            # Added for results that only include one set (ex. LeagueLeaders)
+            headers = json_inp['resultSet']['headers']
+            values = json_inp['resultSet']['rowSet']
+        return pd.DataFrame(values, columns=headers)
+
+
+def _get_json(endpoint, params, referer='scores'):
+    """
+    Internal method to streamline our requests / json getting
+
+    Args:
+        endpoint (str): endpoint to be called from the API
+        params (dict): parameters to be passed to the API
+
+    Raises:
+        HTTPError: if requests hits a status code != 200
+
+    Returns:
+        json (json): json object for selected API call
+    """
+    _get = get(BASE_URL.format(endpoint=endpoint), params=params,
+               headers=HEADERS)
+    _get.raise_for_status()
+
+    return _get.json()
